Root hash signature verification
