+++ sigstore/transparency.py
@@ -18,8 +18,6 @@ Transparency log data structures.
 
 from __future__ import annotations
 
-import base64
-import re
 from dataclasses import dataclass
 from typing import Any, Dict, List, Optional
 
@@ -119,119 +117,6 @@ class LogEntry:
         }
 
         return encode_canonical(payload).encode()  # type: ignore
-
-
-# FIXME(jl): this does not feel like a de novo definition...
-# does this exist already in sigstore-python (or its depenencices)?
-class Signature(BaseModel):
-    name: StrictStr
-    sig_hash: bytes
-    sig_base64: bytes
-
-
-class Checkpoint(BaseModel):
-    origin: StrictStr
-    log_size: int
-    log_hash: StrictStr
-    other_content: list[str]
-
-    @classmethod
-    def from_text(cls, text: str) -> Checkpoint:
-        """
-        Serialize from the text header ("note") of a SignedNote.
-
-        A checkpoint contains:
-        -
-        """
-
-        lines = text.strip().split("\n")
-        if len(lines) < 4:
-            raise ValueError("Malformed Checkpoint: too few items in header!")
-
-        origin = lines[0]
-        if len(origin) == 0:
-            raise ValueError("Malformed Checkpoint: empty origin!")
-
-        log_size = int(lines[1])
-        root_hash = base64.b64decode(lines[2]).hex()
-
-        return Checkpoint(
-            origin=origin,
-            log_size=log_size,
-            log_hash=root_hash,
-            other_content=lines[3:],
-        )
-
-
-class SignedNote(BaseModel):
-    note: StrictStr = Field(..., alias="note")
-    signatures: list[Signature] = Field(..., alias="signatures")
-
-    @validator("signatures")
-    def _signatures_nonempty(cls, v: List[bytes]) -> List[bytes]:
-        if len(v) < 1:
-            raise ValueError("Inclusion proof signatures list is empty!")
-        return v
-
-    @classmethod
-    def from_text(cls, text: str) -> SignedNote:
-        """
-        Serialize from a bundled text 'note'.
-
-        A note contains:
-        - a name, a string associated with the signer,
-        - a separator blank line,
-        - and signature(s), each signature takes the form
-            `\u2014 NAME SIGNATURE\n`
-          (where \u2014 == em dash).
-
-        An adaptation of the Rekor's `UnmarshalText`:
-        https://github.com/sigstore/rekor/blob/4b1fa6661cc6dfbc844b4c6ed9b1f44e7c5ae1c0/pkg/util/signed_note.go#L141
-        """
-
-        separator: str = "\n\n"
-        if text.count(separator) != 1:
-            raise ValueError(
-                "Note must contain one blank line, deliniating the text from the signature block"
-            )
-        split = text.index(separator)
-
-        header: str = text[: split + 1]
-        data: str = text[split + len(separator) :]
-
-        if len(data) == 0:
-            raise ValueError("Malformed Note: must contain at least one signature!")
-        if data[-1] != "\n":
-            raise ValueError("Malformed Note: data section must end with newline!")
-
-        signatures: list[Signature] = []
-
-        sig_parser = re.compile(r"\u2014 (\S+) (\S+)\n")
-        for (name, signature) in re.findall(sig_parser, data):
-            signature_bytes: bytes = base64.b64decode(signature)
-            if len(signature_bytes) < 5:
-                raise ValueError("Malformed Note: signature contains too few bytes")
-
-            signature = Signature(
-                name=name,
-                # FIXME(jl): In Go, construct an big-endian UInt32 from 4 bytes. Is this equivalent?
-                sig_hash=signature_bytes[0:4],
-                sig_base64=base64.b64encode(signature_bytes[4:]),
-            )
-            signatures.append(signature)
-
-        return cls(note=header, signatures=signatures)
-
-
-class SignedCheckpoint(BaseModel):
-    signed_note: SignedNote
-    checkpoint: Checkpoint
-
-    @classmethod
-    def from_text(cls, text: str) -> SignedCheckpoint:
-        signed_note = SignedNote.from_text(text)
-        checkpoint = Checkpoint.from_text(signed_note.note)
-        return cls(signed_note=signed_note, checkpoint=checkpoint)
 
 
 class LogInclusionProof(BaseModel):
